feat(internship): Add student-specific queries for internship applications and documents

+++ backend/app/controllers/internship.py
@@ -1,15 +1,13 @@
 from sqlalchemy.orm import Session
-from fastapi import HTTPException, status
 from models.internship import Internship, InternshipApplication, InternshipDocument
 from repos.internship import InternshipRepo, InternshipApplicationRepo, InternshipDocumentRepo
 from schemas.internship import (
-    InternshipInput, 
-    InternshipApplicationInput,
-    InternshipApplicationUpdate, 
-    InternshipDocumentInput,
-    InternshipDocumentUpdate,
+    InternshipInput, InternshipUpdate,
+    InternshipApplicationInput, InternshipApplicationUpdate, 
+    InternshipDocumentInput, InternshipDocumentUpdate,
 )
 from utils.utils import orm_to_dict, map_to_model
+
 
 # ---------------------- INTERNSHIP ----------------------
 
@@ -52,7 +50,7 @@ def get_internships_by_site(site_id: int, db: Session):
     internships_response = [orm_to_dict(internship) for internship in internships]
     return internships_response
 
-def update_internship(internship_id: int, internship_input: InternshipInput, db: Session):
+def update_internship(internship_id: int, internship_input: InternshipUpdate, db: Session):
     update_data = internship_input.dict(exclude_unset=True)
     internship_repo = InternshipRepo(db)
     updated_internship = internship_repo.update(internship_id, update_data)
@@ -90,6 +88,14 @@ def get_internship_application(application_id: int, db: Session):
     internship_application_response = orm_to_dict(internship_application)
     return internship_application_response
 
+def get_internship_applications_by_student(student_id: int, db: Session):
+    internship_application_repo = InternshipApplicationRepo(db)
+    internship_applications = internship_application_repo.get_by_student_id(student_id)
+    if not internship_applications:
+        return []
+    internship_applications_response = [orm_to_dict(internship_application) for internship_application in internship_applications]
+    return internship_applications_response
+
 def update_internship_application(application_id: int, internship_application_input: InternshipApplicationUpdate, db: Session):
     update_data = internship_application_input.dict(exclude_unset=True)
     internship_application_repo = InternshipApplicationRepo(db)
@@ -112,39 +118,37 @@ def create_internship_document(internship_document: InternshipDocumentInput, db:
     internship_document_response = orm_to_dict(created_internship_document)
     return internship_document_response
 
-def get_all_internship_documents(db: Session):
+def get_all_internship_documents(internship_id: int, db: Session):
     internship_document_repo = InternshipDocumentRepo(db)
-    internship_documents = internship_document_repo.get_all()
+    internship_documents = internship_document_repo.get_all(internship_id)
     if not internship_documents:
         return []
     internship_documents_response = [orm_to_dict(internship_document) for internship_document in internship_documents]
     return internship_documents_response
 
-def get_internship_documents_by_id(document_id: str, db: Session):
+def get_internship_documents_by_id(internship_id: int, document_id: int, db: Session):
     internship_document_repo = InternshipDocumentRepo(db)
-    internship_document = internship_document_repo.get_by_id(document_id)
+    internship_document = internship_document_repo.get_by_id(internship_id, document_id)
     if not internship_document:
         return None
     internship_document_response = orm_to_dict(internship_document)
     return internship_document_response
 
-def get_internship_documents_by_internship(internship_id: int, db: Session):
-    internship_document_repo = InternshipDocumentRepo(db)
-    internship_documents = internship_document_repo.get_by_internship_id(internship_id)
-    if not internship_documents:
-        return []
-    internship_documents_response = [orm_to_dict(internship_document) for internship_document in internship_documents]
-    return internship_documents_response
-
-def update_internship_document(document_id: str, internship_document_input: InternshipDocumentUpdate, db: Session):
+def update_internship_document(internship_id: int, document_id: int, internship_document_input: InternshipDocumentUpdate, db: Session):
     update_data = internship_document_input.dict(exclude_unset=True)
     internship_document_repo = InternshipDocumentRepo(db)
+    document = internship_document_repo.get_by_id(internship_id, document_id)
+    if not document:
+        return None
     updated_internship_document = internship_document_repo.update(document_id, update_data)
     if not updated_internship_document:
         return None
     updated_internship_document_response = orm_to_dict(updated_internship_document)
     return updated_internship_document_response
 
-def delete_internship_document(document_id: str, db: Session):
+def delete_internship_document(internship_id: int, document_id: int, db: Session):
     internship_document_repo = InternshipDocumentRepo(db)
+    document = internship_document_repo.get_by_id(internship_id, document_id)
+    if not document:
+        return False
     return internship_document_repo.delete(document_id)
